+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-12-14
+
+### Added
+
+- Added the ability to enable HSTS in custom server fragments configurations, to enable it you will need to add in the server definition the `#hstsheader` annotation
+
 ## 2022-12-12
 
 ### Added